Move OctMatrix and OctValue to SmtLibUtils
