+++ trunk/source/AbstractInterpretationV2/src/de/uni_freiburg/informatik/ultimate/plugins/analysis/abstractinterpretationv2/domain/relational/octagon/OctAssumeProcessor.java
@@ -52,6 +52,8 @@ import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.absint.IAbstrac
 import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.boogie.IBoogieSymbolTableVariableProvider;
 import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.cfg.structure.IcfgEdge;
 import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.cfg.variables.IProgramVarOrConst;
+import de.uni_freiburg.informatik.ultimate.lib.smtlibutils.octagon.NumericUtils;
+import de.uni_freiburg.informatik.ultimate.lib.smtlibutils.octagon.OctValue;
 import de.uni_freiburg.informatik.ultimate.plugins.analysis.abstractinterpretationv2.domain.nonrelational.interval.IntervalDomainState;
 import de.uni_freiburg.informatik.ultimate.plugins.analysis.abstractinterpretationv2.domain.nonrelational.interval.IntervalPostOperator;
 import de.uni_freiburg.informatik.ultimate.plugins.analysis.abstractinterpretationv2.domain.util.typeutils.TypeUtils;
@@ -402,7 +404,7 @@ public class OctAssumeProcessor {
 		leC = geC = affExpr.getConstant().negate();
 		if (intRelation) {
 			// in case of integers: (affExpr - c != 0) as (affExpr <= c-1) or (affExpr >= c+1)
-			if (AbsIntUtil.isIntegral(leC)) {
+			if (NumericUtils.isIntegral(leC)) {
 				leC = leC.subtract(BigDecimal.ONE);
 				geC = geC.add(BigDecimal.ONE);
 			} else {
@@ -489,7 +491,7 @@ public class OctAssumeProcessor {
 
 		// from now on handle (affExpr - c == 0) as (affExpr == c) ----------------
 		final BigDecimal c = affExpr.getConstant().negate();
-		if (intRelation && !AbsIntUtil.isIntegral(c)) {
+		if (intRelation && !NumericUtils.isIntegral(c)) {
 			// (assume intVar == 1.5) is equivalent to (assume false).
 			return new ArrayList<>();
 		}
@@ -608,7 +610,7 @@ public class OctAssumeProcessor {
 		// from now on handle (affExpr - c <= 0) as (affExpr <= c) ----------------
 		BigDecimal c = affExpr.getConstant().negate();
 		if (intRelation) {
-			if (!AbsIntUtil.isIntegral(c)) {
+			if (!NumericUtils.isIntegral(c)) {
 				// int <= 1.5 ==> int <= 1
 				// int < 1.5 ==> int <= 1
 				c = c.setScale(0, RoundingMode.FLOOR);
