+++ trunk/source/AbstractInterpretationV2/src/de/uni_freiburg/informatik/ultimate/plugins/analysis/abstractinterpretationv2/domain/relational/octagon/OctExponentialWideningOperator.java
@@ -31,6 +31,8 @@ import java.math.BigDecimal;
 import java.util.function.BiFunction;
 
 import de.uni_freiburg.informatik.ultimate.lib.modelcheckerutils.absint.IAbstractStateBinaryOperator;
+import de.uni_freiburg.informatik.ultimate.lib.smtlibutils.octagon.OctMatrix;
+import de.uni_freiburg.informatik.ultimate.lib.smtlibutils.octagon.OctValue;
 
 /**
  * Widening operator for octagons that increases values a binary exponential backoff.
