+++ trunk/source/AbstractInterpretationV2Test/src/de/uni_freiburg/informatik/ultimate/plugins/analysis/abstractinterpretationv2/domain/relational/octagon/ClosureBenchmark.java
@@ -13,6 +13,8 @@ import java.util.List;
 import java.util.function.Consumer;
 import java.util.stream.Collectors;
 
+import de.uni_freiburg.informatik.ultimate.lib.smtlibutils.octagon.OctMatrix;
+
 /**
  * Measures speed of different closures on matrices from Benchmark set.
  *
