+++ trunk/source/Library-SmtLibUtils/src/de/uni_freiburg/informatik/ultimate/lib/smtlibutils/octagon/OctValue.java
@@ -25,15 +25,13 @@
  * to convey the resulting work.
  */
 
-package de.uni_freiburg.informatik.ultimate.plugins.analysis.abstractinterpretationv2.domain.relational.octagon;
+package de.uni_freiburg.informatik.ultimate.lib.smtlibutils.octagon;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 
 import de.uni_freiburg.informatik.ultimate.lib.smtlibutils.SmtUtils;
 import de.uni_freiburg.informatik.ultimate.logic.Rational;
-import de.uni_freiburg.informatik.ultimate.plugins.analysis.abstractinterpretationv2.domain.nonrelational.interval.IntervalValue;
-import de.uni_freiburg.informatik.ultimate.plugins.analysis.abstractinterpretationv2.util.AbsIntUtil;
 
 /**
  * Values for {@link OctMatrix} entries.
@@ -57,16 +55,6 @@ public class OctValue implements Comparable<OctValue> {
 	/** Creates a new OctValue with value infinity. */
 	private OctValue() {
 		// mValue is already null => represents infinity
-	}
-
-	/**
-	 * Creates a new OctValue from an {@link IntervalValue}.
-	 *
-	 * @param ivlValue
-	 *            value
-	 */
-	public OctValue(final IntervalValue ivlValue) {
-		mValue = ivlValue.isInfinity() ? null : ivlValue.getValue();
 	}
 
 	/**
@@ -102,16 +90,7 @@ public class OctValue implements Comparable<OctValue> {
 		if ("inf".equals(s)) {
 			return INFINITY;
 		}
-		return new OctValue(AbsIntUtil.sanitizeBigDecimalValue(s));
-	}
-
-	/**
-	 * Converts this OctValue into an {@link IntervalValue}.
-	 *
-	 * @return IntervalValue
-	 */
-	public IntervalValue toIvlValue() {
-		return mValue == null ? new IntervalValue() : new IntervalValue(mValue);
+		return new OctValue(NumericUtils.sanitizeBigDecimalValue(s));
 	}
 
 	/** @return This value is infinity */
@@ -134,7 +113,7 @@ public class OctValue implements Comparable<OctValue> {
 		}
 		return SmtUtils.toRational(mValue);
 	}
-	
+
 	/**
 	 * Calculates the sum of this and another OctValue. The sum of infinity and something other is infinity.
 	 *
